Lemonade Stand game in Python 3
The game uses Python's curses module to render the game visually. The game uses classes, but the current list of classes is to be revised as it contains unnecessary classes and needs a few additional classes. I will continue to work on this game on my spare time despite what my instructor had to say in this regard. Just want to make it work. :)